OXDEV-8251 Move integration tests

+++ source/Core/Decryptor.php
@@ -43,7 +43,7 @@ class Decryptor
     protected function formKey($key, $string)
     {
         $key = '_' . $key;
-        $keyLength = (strlen($string) / strlen($key)) + 5;
+        $keyLength = (int) (strlen($string) / strlen($key)) + 5;
 
         return str_repeat($key, $keyLength);
     }
